Moved booking logic from views.py to new file
