+++ gsr_booking/views.py
@@ -1,8 +1,3 @@
-import requests
-import datetime
-import math
-import random
-
 from django.contrib.auth import get_user_model
 from django.db.models import Prefetch, Q
 from django.http import HttpResponseForbidden
@@ -24,6 +19,7 @@ from gsr_booking.serializers import (
     GroupBookingRequestSerializer,
 )
 
+from .booking_logic import BookingLogic
 
 User = get_user_model()
 
@@ -279,7 +275,8 @@ class GroupViewSet(viewsets.ModelViewSet):
         if not group.has_member(request.user) or not group.is_admin(request.user):
             return HttpResponseForbidden()
 
-        result_json = self.book_room_for_group(group, 
+        booking_logic = BookingLogic()
+        result_json = booking_logic.book_room_for_group(group, 
             booking_data['is_wharton'], 
             booking_data['room'], 
             booking_data['lid'], 
@@ -288,280 +285,3 @@ class GroupViewSet(viewsets.ModelViewSet):
             request.user.username)
 
         return Response(result_json)
-
-    DATE_FORMAT_STR = "%Y-%m-%dT%H:%M:%S%z"
-    MAX_SLOT_HRS = 2.0  # the longest booking allowed per person
-    MIN_SLOT_HRS = 0.5  # the minimum booking allowed per person
-
-    def book_room_for_group(self, group, is_wharton, room, lid, start, end, requester_pennkey):
-        # makes a request to labs api server to book rooms, and returns result json if successful
-        members = group.get_pennkey_active_members()
-        if len(members) < 1:
-            return {
-                "complete_success": False,
-                "partial_success": False,
-                "error": "No members in the group have enabled their pennkey to be used for group bookings"
-            }
-
-        #randomize order, and put requester first in the order
-        random.shuffle(members)
-        for (i, member) in enumerate(members): #puts the requester as the first person
-            if (member['username'] == requester_pennkey):
-                temp = members[0]
-                members[0] = member
-                members[i] = temp
-                break
-
-        if is_wharton:  # huntsman reservation
-            return {"complete_success": False, "partial_success": False, "error": "Unable to book huntsman rooms yet"}
-        else:  # lib reservation
-           
-            #Find the first timeslot to book for (next_start, next_end)
-            START_DATE = datetime.datetime.strptime(start, self.DATE_FORMAT_STR)
-            END_DATE = datetime.datetime.strptime(end, self.DATE_FORMAT_STR)
-            next_start = START_DATE
-            next_end = min(END_DATE, next_start + datetime.timedelta(hours=self.MAX_SLOT_HRS))
-
-            # loop through each member, and attempt to book on their behalf
-            bookings = {}
-            failed_members = []  #store the members w/ failed bookings in here
-
-            for member in members:
-                if next_end - next_start < datetime.timedelta(hours=0.1):
-                    print("BOOKED EVERYTHING ALREADY")
-                    break
-                if member['user__email'] is "":
-                    print("*User " + member['username'] + " had an empty email")
-                    continue
-
-                # make request to labs-api-server
-                success, error = self.book_room_for_user(
-                    room,
-                    lid,
-                    next_start.strftime(self.DATE_FORMAT_STR),
-                    next_end.strftime(self.DATE_FORMAT_STR),
-                    member['user__email'],
-                )
-                if success:
-                    key = f'{lid}_{room}' 
-                    if not key in bookings:
-                        bookings[key] = []
-                    new_bookings = self.split_booking(next_start, next_end, member['username'], True)
-                    bookings[key].extend(new_bookings)
-                    
-                    next_start = next_end
-                    next_end = min(END_DATE, next_start + datetime.timedelta(hours=self.MAX_SLOT_HRS))
-                    print("Succeeded!")
-                elif self.is_fatal_error(error):
-                    result_json = self.construct_bookings_json_obj(bookings, lid, room, END_DATE, next_start, next_end, error)
-                    return result_json
-                else:
-                    failed_members.append(member)
-                    print("Failed :(")
-
-            print("*Attempting to rebook failed slots now (if any)")
-
-            # if unbooked slots still remain and not all booking requests succeeded, loop through each member again
-            # but get reservation credits first to see how much we can book
-            
-            for member in failed_members:
-                if next_end - next_start < datetime.timedelta(hours=0.1):
-                    print("BOOKED EVERYTHING ALREADY")
-                    break
-                if member['user__email'] is "":
-                    print("empty email address")
-                    continue
-
-                # calculate number of credits already used via getReservations
-                (success, used_credit_hours) = self.get_used_booking_credit_for_user(lid, member['user__email'])
-                remaining_credit_hours = self.MAX_SLOT_HRS - used_credit_hours
-                rounded_remaining_credit_hours = math.floor(2 * remaining_credit_hours) / 2
-                print(f"Found ", rounded_remaining_credit_hours, " remaining booking hrs!")
-                if success and remaining_credit_hours >= self.MIN_SLOT_HRS:
-                    
-                    next_end = min(
-                        END_DATE,
-                        next_start + datetime.timedelta(hours=rounded_remaining_credit_hours),
-                    )
-                    (success, error) = self.book_room_for_user(
-                        room,
-                        lid,
-                        next_start.strftime(self.DATE_FORMAT_STR),
-                        next_end.strftime(self.DATE_FORMAT_STR),
-                        member['user__email'],
-                    )
-                    if success:
-                        key = f'{lid}_{room}' 
-                        if not key in bookings:
-                            bookings[key] = []
-                        new_bookings = self.split_booking(next_start, next_end, member['username'], True)
-                        bookings[key].extend(new_bookings)
-
-                        next_start = next_end
-                        next_end = min(
-                            END_DATE, next_start + datetime.timedelta(hours=MAX_SLOT_HRS)
-                        )
-                        print("Suceeded!")
-                    elif self.is_fatal_error(error):
-                        result_json = self.construct_bookings_json_obj(bookings, lid, room, END_DATE, next_start, next_end, error)
-                        return result_json
-            result_json = self.construct_bookings_json_obj(bookings, lid, room, END_DATE, next_start, next_end, error)
-            return result_json
-    def is_fatal_error(self, error):
-        #consider all errors fatal unless its a daily limit error
-        if error is not None:
-            if ('daily limit'.lower() not in error.lower()):
-                return False
-            return True
-        return False
-
-    def construct_bookings_json_obj(self, succesful_bookings, lid, room, end, next_start, next_end, error):
-        #takes in a set of bookings and constructs a JSON object from them (after pre-processing)
-        bookings = succesful_bookings
-        complete_success = (next_end - next_start < datetime.timedelta(hours=0.1)) and (error is None)
-        partial_success = (len(bookings) > 0)
-        result_json = {
-            "complete_success": complete_success,
-            "partial_success": partial_success,
-            "rooms": []
-        }
-
-        #add failed bookings to bookings
-        failed_bookings = self.split_booking(next_start, end, None, False)
-        failed_key = f'{lid}_{room}' 
-        if not failed_key in bookings:
-            bookings[failed_key] = []
-        failed_split_bookings = self.split_booking(next_start, next_end, None, False)
-        bookings[failed_key].extend(failed_split_bookings)
-
-        #add successful bookings to result_json
-        for (key, bookings_array) in bookings.items():
-            key_split = key.split('_')
-            lid = key_split[0]
-            room = key_split[1]
-            result_json['rooms'].append({
-                'lid': lid,
-                'room': room,
-                'bookings': bookings_array
-            })
-
-        #handle potential error
-        if error is not None:
-            if "not a valid".lower() in error.lower():
-                result_json['error'] = 'Attempted to book for an invalid time slot'
-            elif "daily limit".lower() in error.lower():
-                result_json['error'] = 'Group does not have sufficient booking credits'
-            else:
-                result_json['error'] = str(error)
-        return result_json    
-
-        
-
-    def get_used_booking_credit_for_user(self, lid, email):
-        # returns a user's used booking credit (in hours) for a specific building (lid)
-        RESERVATIONS_URL = "https://api.pennlabs.org/studyspaces/reservations"
-        if lid == "1":
-            return (False, 0)  # doesn't support huntsman yet
-        try:
-            print(f"*Attempting to get reservation credits for {email:s}")
-            r = requests.get(RESERVATIONS_URL + "?email=" + email)
-            if r.status_code == 200:
-                resp_data = r.json()
-                reservations = resp_data["reservations"]
-                used_credit_hours = 0
-                for reservation in reservations:
-                    from_date = datetime.datetime.strptime(reservation["fromDate"], self.DATE_FORMAT_STR)
-                    to_date = datetime.datetime.strptime(reservation["toDate"], self.DATE_FORMAT_STR)
-                    reservation_hours = (to_date - from_date).total_seconds() / 3600
-                    if str(reservation["lid"]) == lid and reservation_hours > 0.1:
-                        used_credit_hours += reservation_hours
-                return (True, used_credit_hours)
-        except requests.exceptions.RequestException as e:
-            print(e)
-        return (False, 0)
-
-    def split_booking(self, start, end, pennkey, booked):
-        # splits a booking into smaller bookings (of min_slot_hrs) for displaying to user
-        bookings = []
-        temp_start = start
-        temp_end = temp_start + datetime.timedelta(hours=self.MIN_SLOT_HRS)
-        while (end - temp_start >= datetime.timedelta(hours=self.MIN_SLOT_HRS)):
-            booking_obj = {
-                'start': temp_start.strftime(self.DATE_FORMAT_STR), 
-                'end': temp_end.strftime(self.DATE_FORMAT_STR), 
-                'booked': booked
-            }
-            if pennkey is not None:
-                booking_obj['pennkey'] = pennkey
-            bookings.append(booking_obj)
-            temp_start = temp_end
-            temp_end += datetime.timedelta(hours=self.MIN_SLOT_HRS)
-        return bookings
-
-    def book_room_for_user(self, room, lid, start, end, email):
-        # tries to make a booking for an individual user, and returns success or not (and the error) in a tuple
-        if lid is "1":
-            return False  # does not support huntsman booking yet
-        print(f"*Attempting to book room {room:d} from {start:s} to {end:s} via {email:s}")
-        BOOKING_URL = "https://api.pennlabs.org/studyspaces/book"
-        form_data = {
-            "firstname": "Group GSR User",
-            "lastname": "Group GSR User",
-            "groupname": "Group GSR",
-            "size": "2-3",
-            "phone": "2158986533",
-            "room": room,
-            "lid": lid,
-            "start": start,
-            "end": end,
-            "email": email,
-        }
-
-        try:
-            r = requests.post(BOOKING_URL, data=form_data)
-            if r.status_code == 200:
-                resp_data = r.json()
-                if "error" in resp_data and resp_data["error"] is not None:
-                    print("error: " + resp_data["error"])
-                    return (False, resp_data["error"])
-                return (resp_data["results"], None)
-        except requests.exceptions.RequestException as e:
-            print("error: " + str(e))
-        return (False, None)
-
-    def make_booking_request_huntsman(self, roomid, startTime, endTime, lid, sessionid):
-        # makes a request to labs api server to book rooms, and returns result json if successful
-        booking_url = "https://api.pennlabs.org/studyspaces/book"
-        if lid == "1":
-            return {"success": False, "error": "Unable to book huntsman rooms yet"}
-        form_data = {
-            "room": roomid,
-            "start": startTime,
-            "end": endTime,
-            "lid": lid,
-            "sessionid": sessionid,
-        }
-
-        # catch potential error in request
-        result = {}
-        try:
-            r = requests.post(booking_url, data=form_data)
-        except requests.exceptions.RequestException as e:
-            print("error: " + str(e))
-            result["error"] = str(e)
-            result["success"] = False
-
-        # go through all of them, do it in 90 minute slots. if it fails, see if anyone has
-
-        # construct result json
-        if "r" in locals() and r.status_code == 200:
-            resp_data = r.json()
-            print(resp_data)
-            result["success"] = resp_data["results"]
-            if "error" in resp_data:
-                result["error"] = resp_data["error"]
-        else:
-            result["success"] = False
-            result["error"] = "Call to labs-api-server failed"
-
-        return result
